feat: Generate a copy of json_fields before adding message
We had a problem where we wanted to log a payload to another service via "json_fields", however, the logging handler added the key "message" to the payload so we started having errors. I propose to generate a copy of the json_fields instead of using the user supplied one.