fix: make atexit registration conditional on background transport
The background transport fails on Python 3.12 when working with gRPC in case
there are messages in the queue when the Python interpreter exits. Making the
registration of the atexit callback conditional so that users can circumvent this
error by explicitly stopping the logging when their application terminates.

We can consider making the default value depend on the Python version at
runtime, but it was decided to not change the default behaviour, regardless
of the Python version.

Fixes #850 and #855 (if users would setup the logging disabling the atexit
hook)